refactor: extract Scopes Session() warning to its own file
Phase 3 (warning when Session/WithContext/Debug is called inside a
Scopes callback) is independent of the rest of the analyzer:

  - It is a temporary workaround for go-gorm/gorm#7592.
  - It has a known sunset (delete once the upstream fix ships in a
    tagged release we still support).
  - It does not interact with pollution tracking or pure-function
    validation.

Move scopesWarning, validateScopesCallback, isScopesCallback,
storeRefersToFunction, and getMethodName out of internal/analyzer.go
into internal/scopes_session_warning.go, with a header that documents
the GORM bug links and the explicit removal condition. analyzer.go
keeps just the call into validateScopesCallback at the orchestration
site.

No behavioral change.
