+++ internal/analyzer.go
@@ -165,175 +165,6 @@ func RunSSA(
 }
 
 // =============================================================================
-// Scopes Callback Validation
-// =============================================================================
-
-// scopesWarning represents a warning about Session() in Scopes callback.
-type scopesWarning struct {
-	Pos     token.Pos
-	Message string
-}
-
-// validateScopesCallback checks if a closure is a Scopes callback and warns
-// about Session/WithContext/Debug calls which cause transaction leaks in GORM.
-//
-// This is a temporary warning for GORM bug: https://github.com/go-gorm/gorm/issues/7592
-func validateScopesCallback(fn *ssa.Function) []scopesWarning {
-	// Only check closures (functions with parents)
-	parent := fn.Parent()
-	if parent == nil {
-		return nil
-	}
-
-	// Check if this closure is passed to Scopes
-	if !isScopesCallback(fn, parent) {
-		return nil
-	}
-
-	// Scan for Session/WithContext/Debug calls
-	var warnings []scopesWarning
-	for _, block := range fn.Blocks {
-		for _, instr := range block.Instrs {
-			call, ok := instr.(*ssa.Call)
-			if !ok {
-				continue
-			}
-
-			methodName := getMethodName(call)
-			if methodName == "" {
-				continue
-			}
-
-			// Check for immutable-returning methods that cause transaction leak in Scopes
-			switch methodName {
-			case "Session":
-				warnings = append(warnings, scopesWarning{
-					Pos:     call.Pos(),
-					Message: "Session() in Scopes callback causes transaction leak (GORM bug)",
-				})
-			case "WithContext":
-				warnings = append(warnings, scopesWarning{
-					Pos:     call.Pos(),
-					Message: "WithContext() in Scopes callback causes transaction leak (calls Session internally)",
-				})
-			case "Debug":
-				warnings = append(warnings, scopesWarning{
-					Pos:     call.Pos(),
-					Message: "Debug() in Scopes callback causes transaction leak (calls Session internally)",
-				})
-			}
-		}
-	}
-
-	return warnings
-}
-
-// storeRefersToFunction reports whether store's value is fn directly or
-// a MakeClosure wrapping fn.
-func storeRefersToFunction(store *ssa.Store, fn *ssa.Function) bool {
-	if directFn, ok := store.Val.(*ssa.Function); ok && directFn == fn {
-		return true
-	}
-	if mc, ok := store.Val.(*ssa.MakeClosure); ok {
-		if closureFn, ok := mc.Fn.(*ssa.Function); ok && closureFn == fn {
-			return true
-		}
-	}
-	return false
-}
-
-// isScopesCallback checks if fn is a callback passed to the Scopes method.
-func isScopesCallback(fn *ssa.Function, parent *ssa.Function) bool {
-	// Look for Store instructions that store our function
-	for _, block := range parent.Blocks {
-		for _, instr := range block.Instrs {
-			store, ok := instr.(*ssa.Store)
-			if !ok || !storeRefersToFunction(store, fn) {
-				continue
-			}
-
-			// Found a Store of our function. Trace the destination.
-			// store.Addr is typically an IndexAddr for variadic calls
-			indexAddr, ok := store.Addr.(*ssa.IndexAddr)
-			if !ok {
-				continue
-			}
-
-			// indexAddr.X is the array being indexed
-			alloc, ok := indexAddr.X.(*ssa.Alloc)
-			if !ok {
-				continue
-			}
-
-			// Check if this array is used in a Scopes call
-			allocRefs := alloc.Referrers()
-			if allocRefs == nil {
-				continue
-			}
-			for _, ref := range *allocRefs {
-				// Look for Slice that converts array to slice
-				if slice, ok := ref.(*ssa.Slice); ok {
-					sliceRefs := slice.Referrers()
-					if sliceRefs == nil {
-						continue
-					}
-					for _, sliceRef := range *sliceRefs {
-						if call, ok := sliceRef.(*ssa.Call); ok {
-							if getMethodName(call) == "Scopes" {
-								return true
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	// Also check for direct function reference (no free variables, single arg)
-	for _, block := range parent.Blocks {
-		for _, instr := range block.Instrs {
-			call, ok := instr.(*ssa.Call)
-			if !ok {
-				continue
-			}
-			if getMethodName(call) != "Scopes" {
-				continue
-			}
-			// Check all arguments
-			for _, arg := range call.Call.Args {
-				// Direct function reference
-				if funcVal, ok := arg.(*ssa.Function); ok && funcVal == fn {
-					return true
-				}
-				// MakeClosure
-				if mc, ok := arg.(*ssa.MakeClosure); ok {
-					if closureFn, ok := mc.Fn.(*ssa.Function); ok && closureFn == fn {
-						return true
-					}
-				}
-			}
-		}
-	}
-
-	return false
-}
-
-// getMethodName extracts the method name from a call instruction.
-func getMethodName(call *ssa.Call) string {
-	// Interface method call
-	if call.Call.IsInvoke() {
-		return call.Call.Method.Name()
-	}
-
-	// Static callee
-	if callee := call.Call.StaticCallee(); callee != nil {
-		return callee.Name()
-	}
-
-	return ""
-}
-
-// =============================================================================
 // SSA Checker
 // =============================================================================
